feat: add validations for existent id

+++ MinimalApi/Modules/User/Handlers/DeleteUser.cs
@@ -6,6 +6,11 @@ public static class DeleteUser
 {
     public static async Task<IResult> Handle(int id, IUserRepository userRepository)
     {
+        var exists = await userRepository.Exists(id);
+        if (!exists)
+        {
+            return Results.BadRequest(new { Message = "User not found" });
+        }
         await userRepository.Delete(new Data.Models.User { Id = id });
         return Results.Ok("User deleted successfully");
     }
